+++ docs_manual/conf.py
@@ -12,7 +12,11 @@
 import os
 import sys
 
-sys.path.insert(0, os.path.abspath("."))
+import django
+
+sys.path.insert(0, os.path.abspath("../"))
+os.environ["DJANGO_SETTINGS_MODULE"] = "config.settings.local"
+django.setup()
 
 # Get the project root dir, which is the parent dir of this
 cwd = os.getcwd()
@@ -37,7 +41,7 @@ import varfish  # noqa: E402
 
 # Add any Sphinx extension module names here, as strings. They can be extensions
 # coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
-extensions = []
+extensions = ["sphinx.ext.autodoc", "sphinx.ext.viewcode", "sphinx-jsonschema"]
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]